Split source files into liberty and characterizer
