+++ characterizer/README.md
@@ -1,0 +1,1 @@
+This folder contains python tools for characterizing standard cells (provided as spice sources) and generating liberty file contents from the characterization results.
