+++ liberty/README.md
@@ -1,0 +1,3 @@
+This folder contains python tools for interfacing with liberty files.
+
+Eventually this should probably evolve into a library for reading/writing liberty files, but for our use cases this is sufficient so far.
